feat(ERC721): Added etherscan link for asset-type documents
Originally worked with a few more details on the asset directly in the page, but realised a lot of it was superfluous since no one's going to actually read the hex addresses and we don't have a mapping of hex addresses to metadata.

Have reduced it to simply an external link to Etherscan so that asset owners can manipulate the asset over there instead.